requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
